tst_QGraphicsGridLayout: fix memleaks in setGeometry()
The test function created a parent-less QGraphicsWidget, and while it
added a bunch of child objects to it, that initial graphics-widgets
never got an owner so it and all the children were leaked.

Fix by holding it in a unique_ptr. That is the minimally-invasive fix.

Amends the start of the public history.

Pick-to: 6.6 6.5 5.15
Change-Id: I3ab9db721e1343f2bd8d9fdd0ee6e86733a66da2
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 2e0988190febe65b6d12b7956bd22b04843e2e6d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>