QFileInfoGatherer: Prevent infinite loop
If the path to be added to the queue already exists at position 0
of the queue, loc is set to 0. This causes the next invocation of
lastIndexOf to start at index -1 (the end of the queue). It will
then again find the path at index 0. This means that the loop
will never terminate, causing the application to freeze here.

This patch fixes that by breaking out of the loop if loc is 0.

Pick-to: 6.9
Change-Id: Ib84f0b2766aa39c87f66b94a2e3fbc6ede687521
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit a43b6dfd5eadbdcf46451a120bd2994a5f5bacc3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>