qmake: SBOM 5.15 qmake2cmake parsing fixes
Adjust the syntax of a few project files to ensure qmake2cmake can
parse the files without patches or complicated workarounds.

An accidental incorrect assignment to SOURCES is removed in
freetype.pro, the ftdebug.c is already assigned elsewhere.
qmake treated it as a false condition, basically a no-op.

A few comments are moved out-of-line from their respective
assignments. This avoids issues trying to parse comments in a
multi-line assignment.

Change-Id: I16f19abc7137502f0eb85f2c66d284d8569e69d3
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>