QMetaObjectPrivate: simplify parameterTypeNamesFromSignature()
Take by QByteArrayView instead of `const char *`, this method is only
called on a QByteArray so the size is already known. Use the length to
check for the end of iteration instead of checking for the closing ')'
in multiple places in the while-loops' conditions.

Change-Id: I272231293d07ae699083d2a38898ea2d3cf0e7a0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 8635b9c2955fe80add75e982647cd3e7883347fa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>