macOS: use correct icon pixmaps in native menus
When creating an NSImage from a QIcon to show the icon of the action in
native menus, then we so far didn't take the state (on/off) or mode
(normal/disabled) of the UI element into account. This resulted in a
menu item showing the Normal/Off icon of an action, even if the
underlying action was disabled or checked.

Icons in menus are not common on macOS, but since we have the
information we need in QCocoaMenuItem, we might just as well use it to
get the right pixmap from the QIcon (even if that pixmap is generated
from the Normal/Off pixmap).

Add an overload to our NSImage::imageFromQIcon helpers that takes a mode
and state, call that from the existing overload, and use it in
QCocoaMenuItem::sync.

Pick-to: 6.8
Fixes: QTBUG-135934
Change-Id: Ifce73daf49ce34c874fa5bf652d84b3102a06709
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 84da5450c342655a7e426c5bb07d9accf7a222a1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>