macOS: Make sure window geometry is correct if it was adjusted by the OS
When using setGeometry on a window, the geometry we requested might not
be accepted by the OS. Specifically on macOS, if the change was
completely rejected, no geometry update will be sent from the OS which
would normally adjust our own geometry data in this case. This means we
have to make sure we don't update any internals with the wrong geometry
when this happens case.

One example of when window size changes can be completely rejected is if
we try to move a window outside the top of the screen, which is not
allowed.

Fixes: QTBUG-136045
Pick-to: 6.11
Change-Id: Idd41fba68850a96a7eac24901d0e0b112420f66b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>