QIcon/QPixmapIconEngine: render svgs with the given size
Currently scalable images are rendered with the default size set in the
image and not the size they were added with QIcon::addPixmap(). This
might give scaling artifacts when the both sizes do not match at all.
Therefore set the wanted size through QImageReader::setScaledSize() when
(and only when) the reader supports this. Setting it unconditionally
will also force the scaling of pixel-based formats which is not what we
want.

Task-number: QTBUG-135652
Change-Id: I2dc78587854b4c7c32438e0386385e980a9c4059
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 9eddf930a274757c91290d56d10177f4ff94c4bb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>