Delay script-incompatible language-only translation to end
Inspired by commit 0278a80d6876f38bb37f1d504e6b24a3512efb7a,
"Correct initialization of justAfter in QLocale::uiLanguages()",
but the code that fixes isn't even present on 6.8, so we have no hope
of fixing the issue it addressed in uiLanguages(). However, we can fix
it in QTranslator, so this does.

This radically adapts the commit it is picked from, adding an adapted
form of its pa-Arab-GB test-case for uiLanguages (which, in 6.8, isn't
where the issue surfaces) and devising a way to test, in QTranslator,
the equivalent of what the original was checking for. I shall adapt
that test, in due course, to 6.9 as an amendment to what lead to this
"pick".

As a drive-by, fixed a regex in the QTranslator test to actually check
that .qm is at the end, when present, rather than matching *.qml files
(as it would previously have done).

Task-number: QTBUG-131894
Change-Id: I75ab27153133d2608cf57f0bc998c1e3d0d1432c
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 0278a80d6876f38bb37f1d504e6b24a3512efb7a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit b5b172e34fc1e66097ebfa96e7c5fdec969db2cc)