Windows11Style: make QPushButton a little bit wider
The spacing betwen the border and text in a QPushButton is a little bit
to small because the CE_PushButtonBevel is drawn with 4px less than the
size hint due to the rounded border. Therefore add the 4px for
CT_PushButton.

Pick-to: 6.8
Fixes: QTBUG-126515
Change-Id: Ie64225ad4734399b5da8bc4e73738592f32a7982
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>
(cherry picked from commit d205844e51059c91395cf73ecd04ccc9382ff72c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>