Revert "iOS: Replace variable length array in event dispatcher with QVarLengthArray"
This reverts commit be93c06758260b1e0738ab17a2ecc4ee8f90e87e.

The assumption was that using a QVarLengthArray was equivalent to
the existing C-style VLA, as we initialized the QVarLengthArray
with defaultStackSize as the Prealloc template argument.

However we failed to take into account that we adjust the stack
size via Stack::computeSize(), by adding space for our memory
guard, so we would effectively always end up with the user-main
stack on the heap.

This caused problems for Qt Quick's V4 engine, because it checks
for C++ stack overflow by comparing the current stack pointer to
the base and limit given by pthread_get_stackaddr_np and RLIMIT_STACK,
resulting in "RangeError: Maximum call stack size exceeded." errors
and failure to load the QML.

We can't tweak what pthread_get_stackaddr_np reports for an already
started thread, so in practice we can not put the user-main stack
on the heap, so we revert back to the use of C-style variable length
arrays, which should be fine for this specific use-case.

Change-Id: I05a4bd32697abee71aa9e70f9f5a011cc338f915
Pick-to: 6.8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit ccce91461a39adebde7ec7803b4cc0b97093f77f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>