QStyleSheetStyle: rename initObject() to initWidget()
We can't style QObjects but QWidgets so the name is wrong. Also pass a
QWidget instead a QObject here (and all the callers) - no need to
cast it from QWidget to QObject in the first place and then later cast
it back again to a QWidget.

Change-Id: Ibfeaf97778fc9a5f07cb079b70c48f7f32be93bf
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 836abc860bac3cfc7d08532f0b872741895794d6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>