QByteArray: replace(pos, len,~~): don't detach the underlying data array
If the byte array would detach or reallocate it would copy the data
over, then do the replacements; instead create a new byte array and copy
the data and replacement to it as needed, then swap it with `this`.

The behavior for negative `len` is preserved for historical reasons.

Handle `after`-pointing-into-this by using memmove() which can work with
overlapping ranges.

Task-number: QTBUG-106185
Change-Id: Iedc848bebabf5621f459b11f0edf0e27807b9be0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 0af13fec5a59a25eaf705fa101ebecfe5e0277db)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>