wasm: set focus to m_canvas instead of m_window
Setting focus and contentEditable on m_window causes
innerHTML to build up with characters.

This does not happen when using the m_canvas

The downside is that m_canvas is aria-hidden, but this
should, in principle, not be a problem since
m_canvas should not be focused when a11y is in
effect.

Later aria-hidden might be set only if a11y is in
effect.

It turns out this also fixes tabbing between textfields
and buttons bug.

This is a candidate for manual cherry-picking to
6.9 6.9.1

Task-number: QTBUG-136687
Task-number: QTBUG-136562
Change-Id: I08a9db2c39f9b0b0038c75fd06d3504b736ea031
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>