CRA[network/kernel]: Mark some files critical
The files are parsing or working with data that is obtained from
outside.

tldurl is special in that it is a list we ship or acquire from the
system, but we are working with a string that was obtained from a
user or from the network.

Task-number: QTBUG-135584
Pick-to: 6.9 6.8
Change-Id: Icb13a9acc859e07460f6c10b3f108e65ff130c5c
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit a0aa857fefc00770d1b10efdd6cf6517e81058fe)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>