wasm: rewrite handleKeyForInputContextEvent() logic
Avoid using the "if with fallthrough" pattern. The code
is building up a logical expression in order to determine
if it should send a key event or not, which can be done
more cleanly with boolean variables and sub-expressions.

Change-Id: I2af3501d1c199e1f7f49f8417a21eece219e11ff
Reviewed-by: Lorn Potter <lorn.potter@qt.io>
Reviewed-by: Even Oscar Andersen <even.oscar.andersen@qt.io>
(cherry picked from commit dee0123dc610148b134aa6007cac452ac9f7b206)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>