QUrl: ensure setUrl() / operator=() erase any previous state
QUrlPrivate::parse() did not clear all the state because it was designed
for parsing from the QUrl constructor. The few conditions under which it
retained some memory weren't obvious and weren't tested anywhere (it was
the fragment and query).

This will shed memory from the QStrings if we were the last reference,
only to allocate again. There is some value in attempting to reuse their
buffers by doing resize(0) and then appending in qt_urlRecode(), but
that introduces complexity I didn't want to deal with.

This incidentally fixes the qHash() inequalities too.

Changelog in the previous commit.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-134896
Fixes: QTBUG-134900
Change-Id: I1885d0750ac1109aab61fffdbf7fad775706e61f
Reviewed-by: David Faure <david.faure@kdab.com>
(cherry picked from commit 304a9664d2ebcedd37e9d9094db1d80b2b89fa13)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>