QUrl: use QHashCombine so we use the faster & stronger QString hashing
We were only using the seed for hashing the port number, which is the
weakest of all the uses of the seed, resulting in a weak hash for the
QUrl itself. This commit changes it so QString components of the URL are
themselves also hashed using the seed. As a side benefit, the aeshash()
implementation in qhash.cpp where supported (ARM and x86) is faster than
the siphash() one because it's vectorized.

We're retaining the qHash(QUrl(), seed) == hashing of -1, but it's not
qHash(-1, seed) because QHashCombine adds a constant.

Change-Id: If6466d054fd5a4f05205fffdfbc7b655eae5aefb
Reviewed-by: David Faure <david.faure@kdab.com>
(cherry picked from commit 423278c6ee18a86320d253d8f873347c12f882f5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>