Include showMenuAction() if context menu policy is Qt::ActionsContextMenu
We only report the action if we have any actions to present.

For the other context menu policies we don't know if the user has
overridden contextMenuEvent or connected to customContextMenuRequested,
so we opt to not report the menu action in these cases. A user that
wants to present the menu action will have to subclass QAccessibleWidget,
override actionNames(), and install an accessible interface factory
that returns the specialized QAccessibleWidget.

As of now we do not report QAccessible::ActionChanged anywhere in Qt
when the conditions for actionNames() change, and this includes this
newly added behavior. Initial attempts at plumbing ActionChanged to
the macOS a11y system failed as there is seemingly no way to trigger
an update of this specific state, so this rabbit-hole has been left
for further investigations.

Pick-to: 6.8
Task-number: QTBUG-137126
Change-Id: I538df87603fbcae9ed2d9bcbb16f1de85a08f200
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit bb0c8bbf20fc1cdddd7e171260c4435f360e23d5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 4df9eeb380534bb5687b41f7fde28d587cbd0a64)